refactor: remove dependencies file and moved code to serializer
